Simplify Queue<T> implementations
