+++ rx-java/rxgrpc-stub/src/main/java/com/salesforce/rxgrpc/stub/RxServerStreamObserverAndPublisher.java
@@ -37,9 +37,4 @@ class RxServerStreamObserverAndPublisher<T>
         }
         return QueueFuseable.NONE;
     }
-
-    @Override
-    public boolean offer(T t, T t1) {
-        throw new UnsupportedOperationException();
-    }
 }
